LayoutTests/css3/filters/filter-property-expected.txt

Test the parsing of the -webkit-filter property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS cssRule.type is 1
PASS declaration.length is 1
PASS declaration.getPropertyValue('-webkit-filter') is 'url(#a) url(#b)'
PASS successfullyParsed is true

TEST COMPLETE
